feat: Add reservation guard
