+++ src/app/purchasing-process/(reservation)/select-area/page.tsx
