+++ src/app/purchasing-process/(reservation)/select-seats/page.tsx
